requirements.txt
streamlit==1.35.0
google-generativeai>=0.5.0
boto3==1.34.0
paramiko==3.4.0
python-dotenv==1.0.1
bcrypt==4.1.3
google-auth==2.40.1
google-auth-oauthlib==1.2.2
twilio==9.6.5
jinja2==3.1.4
cryptography==42.0.0
docker==7.1.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
anthropic>=0.28.0
